Completness for quality forms and refactor view
